Moved genQ generators used then in set_run_isr
