+++ test/test_genQ.c
@@ -106,7 +106,7 @@ typedef struct test_genQ
 
 #define BLK_LEN 23
 gStr_t global_blk[BLK_LEN];
-DefineGenQ(globalQ, global_blk, ARR_LEN(global_blk));
+DefineGenQFromStore(globalQ, global_blk, ARR_LEN(global_blk));
 
 void test_global_GenQ_def(void)
 {
@@ -175,37 +175,3 @@ void test_GenQ_Get(void)
     TEST_ASSERT_EQUAL(0, GenQ_HasData(queue));
 }
 
-typedef struct
-{
-    LinkBase_t lnk;
-    char *str;
-    int i;
-} MyThing;
-
-
-void test_stack(void)
-{
-    MyThing items[10];
-    LinkBase_t *q = NULL;
-    
-
-    for (int i=0; i<DIM(items); i++)
-    {
-        items[i].i = i;
-        items[i].str = "Hello";
-        StackPush(&q, &items[i].lnk);
-    }
-
-    int cnt = DIM(items)-1;
-    LinkBase_t *lnk;
-    LinkBase_t *q2;
-    while(lnk = StackPop(&q))
-    {
-        MyThing *item = (MyThing*)lnk;
-        TEST_ASSERT_EQUAL(cnt, item->i);
-        TEST_ASSERT_EQUAL_STRING("Hello", item->str);
-        cnt--;
-        StackPush(&q2, lnk);
-    }
-    TEST_ASSERT_EQUAL_PTR(q2, &items[0] );
-}
